requirements.txt
langchain 
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fitz
pytesseract
# fastapi
# uvicorn
# sse_starlette
# bs4
# pypdf
# chromadb
faiss-cpu
groq
# cassio
# beautifulsoup4
langchain-groq
# wikipedia
# arxiv
# langchainhub
sentence_transformers
PyPDF2
# langchain-objectbox